requirements.txt
Flask==3.0.0
python-telegram-bot==21.9
python-dotenv==1.0.0
